add openid  connect package to client
