+++ src/calendar-backend/docker-compose.yml
@@ -1,8 +1,8 @@
 version: '3.4'
 
 services:
-  calendar-api:
+  cCalendar.api:
     image: ${DOCKER_REGISTRY-}calendarapi
     build:
       context: .
-      dockerfile: calendar-api/Dockerfile
+      dockerfile: Calendar.Api/Dockerfile
